+++ lib/nanoc/base/compilation/compiler.rb
@@ -118,7 +118,6 @@ module Nanoc
       # Preprocess
       self.load_rules
       preprocess
-      site.setup_child_parent_links
       build_reps
       route_reps
 
@@ -165,7 +164,6 @@ module Nanoc
       @stack = []
 
       items.each { |item| item.reps.clear }
-      site.teardown_child_parent_links
 
       site.unload
 
@@ -254,15 +252,15 @@ module Nanoc
           # Get basic path by applying matching rule
           basic_path = rule.apply_to(rep, :compiler => self)
           next if basic_path.nil?
-          if basic_path !~ %r{^/}
+          unless basic_path.to_s.start_with?('/')
             raise RuntimeError, "The path returned for the #{rep.inspect} item representation, “#{basic_path}”, does not start with a slash. Please ensure that all routing rules return a path that starts with a slash."
           end
 
           # Get raw path by prepending output directory
-          rep.raw_paths[snapshot] = @site.config[:output_dir] + basic_path
+          rep.raw_paths[snapshot] = @site.config[:output_dir] + basic_path.to_s
 
           # Get normal path by stripping index filename
-          rep.paths[snapshot] = basic_path
+          rep.paths[snapshot] = basic_path.to_s
           @site.config[:index_filenames].each do |index_filename|
             if rep.paths[snapshot][-index_filename.length..-1] == index_filename
               # Strip and stop