+++ lib/nanoc/base/core_ext/string.rb
@@ -2,6 +2,16 @@
 
 module Nanoc::StringExtensions
 
+  # TODO remove me
+  def stem
+    extension = File.extname(self)
+    if extension.empty?
+      self
+    else
+      self[0..-(1+extension.length)].stem
+    end
+  end
+
   # Calculates the checksum for this string. Any change to this string will
   # result in a different checksum.
   #