+++ lib/nanoc/filters/less.rb
@@ -7,6 +7,18 @@ module Nanoc::Filters
 
     requires 'less'
 
+    def imports_in_content(content)
+      # Find imports
+      imports = []
+      imports.concat(content.scan(/^@import\s+(["'])([^\1]+?)\1;/))
+      imports.concat(content.scan(/^@import\s+url\((["']?)([^)]+?)\1\);/))
+
+      # Append .less to flienames without extension
+      imports.map do |i|
+        i[1].match(/\.(less|css)$/) ? i[1] : i[1] + '.less'
+      end
+    end
+
     # Runs the content through [LESS](http://lesscss.org/).
     # This method takes no options.
     #
@@ -15,38 +27,42 @@ module Nanoc::Filters
     # @return [String] The filtered content
     def run(content, params={})
       # Find imports (hacky)
-      imports = []
-      imports.concat(content.scan(/^@import\s+(["'])([^\1]+?)\1;/))
-      imports.concat(content.scan(/^@import\s+url\((["']?)([^)]+?)\1\);/))
-      imported_filenames = imports.map do |i|
-        i[1].match(/\.(less|css)$/) ? i[1] : i[1] + '.less'
+      imported_filenames = self.imports_in_content(content)
+
+      # FIXME ugly
+      paths = [ Dir.getwd + '/content' ]
+
+      unless imported_filenames.empty?
+        if @item.raw_filename.nil?
+          # FIXME get proper exception
+          raise 'Can only use less filter with items that appear on disk (less limitation)'
+        end
+        current_item_filename = File.join(Dir.getwd, @item.raw_filename)
+        current_dir_filename = File.dirname(current_item_filename)
+
+        paths << current_dir_filename
+
+        # Find items for imported filenames
+        imported_items = imported_filenames.map do |imported_filename|
+          # Find absolute filename for imported item
+          imported_filename_absolute = current_dir_filename + imported_filename
+
+          # Find matching item
+          @items.find do |i|
+            if i.raw_filename.nil?
+              false
+            else
+              this_item_filename = Dir.getwd + i.raw_filename
+              this_item_filename == imported_filename_absolute
+            end
+          end
+        end.compact
+
+        # Create dependencies
+        depend_on(imported_items)
       end
 
-      # Convert to items
-      imported_items = imported_filenames.map do |filename|
-        # Find directory for this item
-        current_dir_pathname = Pathname.new(@item[:content_filename]).dirname.realpath
-
-        # Find absolute pathname for imported item
-        imported_pathname    = Pathname.new(filename)
-        if imported_pathname.relative?
-          imported_pathname = current_dir_pathname + imported_pathname
-        end
-        next if !imported_pathname.exist?
-        imported_filename = imported_pathname.realpath
-
-        # Find matching item
-        @items.find do |i|
-          next if i[:content_filename].nil?
-          Pathname.new(i[:content_filename]).realpath == imported_filename
-        end
-      end.compact
-
-      # Create dependencies
-      depend_on(imported_items)
-
       # Add filename to load path
-      paths = [ File.dirname(@item[:content_filename]) ]
       parser = ::Less::Parser.new(:paths => paths)
       parser.parse(content).to_css params
     end