+++ test/base/test_item.rb
@@ -9,12 +9,6 @@ class Nanoc::ItemTest < Nanoc::TestCase
     assert_equal 'xyz', item.attributes[:abc]
   end
 
-  def test_initialize_with_unclean_identifier
-    item = Nanoc::Item.new("foo", {}, '/foo')
-
-    assert_equal '/foo/', item.identifier.to_s
-  end
-
   def test_frozen_identifier
     item = Nanoc::Item.new("foo", {}, '/foo')
 
@@ -32,7 +26,7 @@ class Nanoc::ItemTest < Nanoc::TestCase
     item = Nanoc::Item.new(
       "content",
       { :one => 'one in item' },
-      '/path/'
+      '/path.md'
     )
 
     # Test finding one
@@ -138,7 +132,7 @@ class Nanoc::ItemTest < Nanoc::TestCase
   end
 
   def test_freeze_should_disallow_changes
-    item = Nanoc::Item.new("foo", { :a => { :b => 123 }}, '/foo/')
+    item = Nanoc::Item.new("foo", { :a => { :b => 123 }}, '/foo')
     item.freeze
 
     assert_raises_frozen_error do
@@ -154,11 +148,11 @@ class Nanoc::ItemTest < Nanoc::TestCase
     item = Nanoc::Item.new(
       "foobar",
       { :a => { :b => 123 }},
-      '/foo/')
+      '/foo')
 
     item = Marshal.load(Marshal.dump(item))
 
-    assert_equal '/foo/', item.identifier.to_s
+    assert_equal '/foo', item.identifier.to_s
     assert_equal 'foobar', item.raw_content
     assert_equal({ :a => { :b => 123 }}, item.attributes)
   end