+++ test/base/test_item_array.rb
@@ -5,8 +5,8 @@ class Nanoc::ItemArrayTest < Nanoc::TestCase
   def setup
     super
 
-    @one = Nanoc::Item.new('Item One', {}, '/one/')
-    @two = Nanoc::Item.new('Item Two', {}, '/two/')
+    @one = Nanoc::Item.new('Item One', {}, '/one.md')
+    @two = Nanoc::Item.new('Item Two', {}, '/two.css')
 
     @items = Nanoc::ItemArray.new
     @items << @one
@@ -14,17 +14,17 @@ class Nanoc::ItemArrayTest < Nanoc::TestCase
   end
 
   def test_change_item_identifier
-    assert_equal @one, @items['/one/']
-    assert_nil @items['/foo/']
+    assert_equal @one, @items['/one.md']
+    assert_nil @items['/foo.txt']
 
-    @one.identifier = '/foo/'
+    @one.identifier = '/foo.txt'
 
-    assert_nil @items['/one/']
-    assert_equal @one, @items['/foo/']
+    assert_nil @items['/one.md']
+    assert_equal @one, @items['/foo.txt']
   end
 
   def test_enumerable
-    assert_equal @one, @items.find { |i| i.identifier == '/one/' }
+    assert_equal @one, @items.find { |i| i.identifier == '/one.md' }
   end
 
   def test_brackets_and_slice_and_at_with_index
@@ -54,13 +54,13 @@ class Nanoc::ItemArrayTest < Nanoc::TestCase
   end
 
   def test_brackets_and_slice_and_at_with_string_identifier
-    assert_equal @one, @items['/one/']
-    assert_equal @one, @items.slice('/one/')
-    assert_equal @one, @items.at('/one/')
+    assert_equal @one, @items['/one.md']
+    assert_equal @one, @items.slice('/one.md')
+    assert_equal @one, @items.at('/one.md')
 
-    assert_equal @two, @items['/two/']
-    assert_equal @two, @items.slice('/two/')
-    assert_equal @two, @items.at('/two/')
+    assert_equal @two, @items['/two.css']
+    assert_equal @two, @items.slice('/two.css')
+    assert_equal @two, @items.at('/two.css')
 
     assert_nil @items['/max-payne/']
     assert_nil @items.slice('/max-payne/')
@@ -68,12 +68,12 @@ class Nanoc::ItemArrayTest < Nanoc::TestCase
   end
 
   def test_brackets_and_slice_and_at_with_object_identifier
-    identifier_one = Nanoc::Identifier.from_string('/one/')
+    identifier_one = Nanoc::Identifier.from_string('/one.md')
     assert_equal @one, @items[identifier_one]
     assert_equal @one, @items.slice(identifier_one)
     assert_equal @one, @items.at(identifier_one)
 
-    identifier_two = Nanoc::Identifier.from_string('/two/')
+    identifier_two = Nanoc::Identifier.from_string('/two.css')
     assert_equal @two, @items[identifier_two]
     assert_equal @two, @items.slice(identifier_two)
     assert_equal @two, @items.at(identifier_two)
@@ -97,17 +97,17 @@ class Nanoc::ItemArrayTest < Nanoc::TestCase
     assert_nil @items.slice('one')
     assert_nil @items.at('one')
 
-    assert_nil @items['//one/']
-    assert_nil @items.slice('//one/')
-    assert_nil @items.at('//one/')
+    assert_nil @items['//one.md']
+    assert_nil @items.slice('//one.md')
+    assert_nil @items.at('//one.md')
   end
 
   def test_brackets_and_slice_and_at_frozen
     @items.freeze
 
-    assert_equal @one, @items['/one/']
-    assert_equal @one, @items.slice('/one/')
-    assert_equal @one, @items.at('/one/')
+    assert_equal @one, @items['/one.md']
+    assert_equal @one, @items.slice('/one.md')
+    assert_equal @one, @items.at('/one.md')
 
     assert_nil @items['/tenthousand/']
     assert_nil @items.slice('/tenthousand/')
@@ -116,32 +116,32 @@ class Nanoc::ItemArrayTest < Nanoc::TestCase
 
   def test_less_than_less_than
     assert_nil @items[2]
-    assert_nil @items['/foo/']
+    assert_nil @items['/foo.txt']
 
-    foo = Nanoc::Item.new('Item Foo', {}, '/foo/')
+    foo = Nanoc::Item.new('Item Foo', {}, '/foo.txt')
     @items << foo
 
     assert_equal foo, @items[2]
-    assert_equal foo, @items['/foo/']
+    assert_equal foo, @items['/foo.txt']
   end
 
   def test_assign
     assert_raises(TypeError) do
-      @items['/blah/'] = Nanoc::Item.new('Item blah', {}, '/blah/')
+      @items['/blah.sass'] = Nanoc::Item.new('Item blah', {}, '/blah.sass')
     end
 
-    new_item =  Nanoc::Item.new('New Item One', {}, '/one-new/')
+    new_item =  Nanoc::Item.new('New Item One', {}, '/one-new.html')
     @items[0] = new_item
 
     assert_equal new_item, @items[0]
-    assert_equal new_item, @items['/one-new/']
-    assert_nil @items['/one/']
+    assert_equal new_item, @items['/one-new.html']
+    assert_nil @items['/one.md']
   end
 
   def test_assign_frozen
     @items.freeze
 
-    new_item = Nanoc::Item.new('New Item One', {}, '/one-new/')
+    new_item = Nanoc::Item.new('New Item One', {}, '/one-new.html')
 
     assert_raises_frozen_error do
       @items[0] = new_item
@@ -155,8 +155,8 @@ class Nanoc::ItemArrayTest < Nanoc::TestCase
     assert_nil @items[1]
     assert_nil @items[2]
 
-    assert_nil @items['/one/']
-    assert_nil @items['/two/']
+    assert_nil @items['/one.md']
+    assert_nil @items['/two.css']
   end
 
   def test_collect_bang
@@ -164,14 +164,14 @@ class Nanoc::ItemArrayTest < Nanoc::TestCase
       Nanoc::Item.new("New #{i.raw_content}", {}, "/new#{i.identifier}")
     end
 
-    assert_nil @items['/one/']
-    assert_nil @items['/two/']
+    assert_nil @items['/one.md']
+    assert_nil @items['/two.css']
 
     assert_equal "New Item One", @items[0].raw_content
-    assert_equal "New Item One", @items['/new/one/'].raw_content
+    assert_equal "New Item One", @items['/new/one.md'].raw_content
 
     assert_equal "New Item Two", @items[1].raw_content
-    assert_equal "New Item Two", @items['/new/two/'].raw_content
+    assert_equal "New Item Two", @items['/new/two.css'].raw_content
   end
 
   def test_collect_bang_frozen
@@ -185,78 +185,78 @@ class Nanoc::ItemArrayTest < Nanoc::TestCase
   end
 
   def test_concat
-    new_item = Nanoc::Item.new('New item', {}, '/new/')
+    new_item = Nanoc::Item.new('New item', {}, '/new.md')
     @items.concat([ new_item ])
 
     assert_equal new_item, @items[2]
-    assert_equal new_item, @items['/new/']
+    assert_equal new_item, @items['/new.md']
   end
 
   def test_delete
     assert_equal @two, @items[1]
-    assert_equal @two, @items['/two/']
+    assert_equal @two, @items['/two.css']
 
     @items.delete(@two)
 
     assert_nil @items[1]
-    assert_nil @items['/two/']
+    assert_nil @items['/two.css']
   end
 
   def test_delete_at
     assert_equal @two, @items[1]
-    assert_equal @two, @items['/two/']
+    assert_equal @two, @items['/two.css']
 
     @items.delete_at(1)
 
     assert_nil @items[1]
-    assert_nil @items['/two/']
+    assert_nil @items['/two.css']
   end
 
   def test_delete_if
     assert_equal @two, @items[1]
-    assert_equal @two, @items['/two/']
+    assert_equal @two, @items['/two.css']
 
-    @items.delete_if { |i| i.identifier == '/two/' }
+    @items.delete_if { |i| i.identifier == '/two.css' }
 
     assert_nil @items[1]
-    assert_nil @items['/two/']
+    assert_nil @items['/two.css']
   end
 
   def test_fill_all
-    @items.fill { |i| Nanoc::Item.new("Item #{i}", {}, "/new/#{i}/") }
+    @items.fill { |i| Nanoc::Item.new("Item #{i}", {}, "/new/#{i}.md") }
 
-    assert_nil @items['/one/']
-    assert_nil @items['/two/']
+    assert_nil @items['/one.md']
+    assert_nil @items['/two.css']
 
     assert_equal "Item 0", @items[0].raw_content
-    assert_equal "Item 0", @items['/new/0/'].raw_content
+    assert_equal "Item 0", @items['/new/0.md'].raw_content
     assert_equal "Item 1", @items[1].raw_content
-    assert_equal "Item 1", @items['/new/1/'].raw_content
+    assert_equal "Item 1", @items['/new/1.md'].raw_content
   end
 
   def test_fill_range
-    @items.fill(1..-1) { |i| Nanoc::Item.new("Item #{i}", {}, "/new/#{i}/") }
+    @items.fill(1..-1) { |i| Nanoc::Item.new("Item #{i}", {}, "/new/#{i}.md") }
 
-    assert_equal @one, @items['/one/']
-    assert_nil @items['/two/']
+    assert_equal @one, @items['/one.md']
+    assert_nil @items['/two.css']
 
     assert_equal @one, @items[0]
-    assert_equal @one, @items['/one/']
+    assert_equal @one, @items['/one.md']
     assert_equal "Item 1", @items[1].raw_content
-    assert_equal "Item 1", @items['/new/1/'].raw_content
+    assert_equal "Item 1", @items['/new/1.md'].raw_content
   end
 
   if Array.new.respond_to?(:keep_if)
     def test_keep_if
       assert_equal @two, @items[1]
-      assert_equal @two, @items['/two/']
+      assert_equal @two, @items['/two.css']
 
-      @items.keep_if { |i| i.identifier == '/one/' }
+      @items.keep_if { |i| i.identifier == '/one.md' }
 
       assert_equal @one, @items[0]
-      assert_equal @one, @items['/one/']
+      assert_equal @one, @items['/one.md']
       assert_nil @items[1]
-      assert_nil @items['/two/']
+      assert_nil @items['/two.css']
     end
   end
 
@@ -264,57 +264,57 @@ class Nanoc::ItemArrayTest < Nanoc::TestCase
     @items.pop
 
     assert_equal @one, @items[0]
-    assert_equal @one, @items['/one/']
+    assert_equal @one, @items['/one.md']
     assert_nil @items[1]
-    assert_nil @items['/two/']
+    assert_nil @items['/two.css']
   end
 
   def test_push
-    pushy = Nanoc::Item.new("Pushy", {}, '/pushy/')
+    pushy = Nanoc::Item.new("Pushy", {}, '/pushy.md')
     @items.push(pushy)
 
     assert_equal @one, @items[0]
-    assert_equal @one, @items['/one/']
+    assert_equal @one, @items['/one.md']
     assert_equal @two, @items[1]
-    assert_equal @two, @items['/two/']
+    assert_equal @two, @items['/two.css']
     assert_equal pushy, @items[2]
-    assert_equal pushy, @items['/pushy/']
+    assert_equal pushy, @items['/pushy.md']
   end
 
   def test_reject_bang
     assert_equal @two, @items[1]
-    assert_equal @two, @items['/two/']
+    assert_equal @two, @items['/two.css']
 
-    @items.reject! { |i| i.identifier == '/two/' }
+    @items.reject! { |i| i.identifier == '/two.css' }
 
     assert_nil @items[1]
-    assert_nil @items['/two/']
+    assert_nil @items['/two.css']
   end
 
   def test_replace
-    max  = Nanoc::Item.new("Max", {}, '/max/')
-    mona = Nanoc::Item.new('Mona', {}, '/mona/')
+    max  = Nanoc::Item.new("Max", {}, '/max.md')
+    mona = Nanoc::Item.new('Mona', {}, '/mona.md')
 
     @items.replace([ max, mona ])
 
-    assert_nil @items['/one/']
-    assert_nil @items['/two/']
+    assert_nil @items['/one.md']
+    assert_nil @items['/two.css']
 
     assert_equal max, @items[0]
-    assert_equal max, @items['/max/']
+    assert_equal max, @items['/max.md']
     assert_equal mona, @items[1]
-    assert_equal mona, @items['/mona/']
+    assert_equal mona, @items['/mona.md']
   end
 
   if Array.new.respond_to?(:select!)
     def test_select_bang
       assert_equal @two, @items[1]
-      assert_equal @two, @items['/two/']
+      assert_equal @two, @items['/two.css']
 
-      @items.select! { |i| i.identifier == '/two/' }
+      @items.select! { |i| i.identifier == '/two.css' }
 
       assert_nil @items[1]
-      assert_nil @items['/one/']
+      assert_nil @items['/one.md']
     end
   end
 
@@ -322,8 +322,8 @@ class Nanoc::ItemArrayTest < Nanoc::TestCase
     @items.shift
 
     assert_equal @two, @items[0]
-    assert_equal @two, @items['/two/']
-    assert_nil @items['/one/']
+    assert_equal @two, @items['/two.css']
+    assert_nil @items['/one.md']
     assert_nil @items[1]
   end
 
@@ -331,21 +331,21 @@ class Nanoc::ItemArrayTest < Nanoc::TestCase
     @items.slice!(1)
 
     assert_equal @one, @items[0]
-    assert_equal @one, @items['/one/']
+    assert_equal @one, @items['/one.md']
     assert_nil @items[1]
-    assert_nil @items['/two/']
+    assert_nil @items['/two.css']
   end
 
   def test_unshift
-    unshifty = Nanoc::Item.new("Unshifty", {}, '/unshifty/')
+    unshifty = Nanoc::Item.new("Unshifty", {}, '/unshifty.md')
     @items.unshift(unshifty)
 
     assert_equal unshifty, @items[0]
-    assert_equal unshifty, @items['/unshifty/']
+    assert_equal unshifty, @items['/unshifty.md']
     assert_equal @one, @items[1]
-    assert_equal @one, @items['/one/']
+    assert_equal @one, @items['/one.md']
     assert_equal @two, @items[2]
-    assert_equal @two, @items['/two/']
+    assert_equal @two, @items['/two.css']
   end
 
 end